backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
PyJWT==2.12.1
sqlparse==0.5.4
tzdata==2025.2
pytest==9.0.3
pytest-django==4.12.0
